% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_to_earth.R
\name{create_contour_kml}
\alias{create_contour_kml}
\title{Create a KML file for the contour overlay}
\usage{
create_contour_kml(
  data,
  png_filename = "contour_overlay.png",
  output_filename = "contour_overlay",
  buffer_percent = 0
)
}
\arguments{
\item{data}{The original data used to create the plot}

\item{png_filename}{The filename of the PNG to use in the KML}

\item{output_filename}{The filename for the KML (without extension)}

\item{buffer_percent}{Optional buffer percentage to add around the data extent}
}
\value{
The filename of the created KML
}
\description{
Create a KML file for the contour overlay
}
