% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_to_earth.R
\name{save_contour_for_overlay}
\alias{save_contour_for_overlay}
\title{Save a contour plot for Google Earth overlay}
\usage{
save_contour_for_overlay(
  plot,
  filename = "contour_overlay",
  width = 10,
  height = 8,
  dpi = 300,
  include_legend = FALSE
)
}
\arguments{
\item{plot}{A ggplot object, typically created with create_contour_plot()}

\item{filename}{The base filename to use (without extension)}

\item{width}{Width of the output image in inches}

\item{height}{Height of the output image in inches}

\item{dpi}{Resolution of the output image}

\item{include_legend}{Whether to include the legend in the export}
}
\value{
The filename of the saved PNG
}
\description{
Save a contour plot for Google Earth overlay
}
